GREEN-964: Moved common classes to java-automation-framework
