+++ testlio-mobile-java-testng-template/src/test/java/com/testlio/constants/FakeTestConstants.java
@@ -1,0 +1,14 @@
+package com.testlio.constants;
+
+public class FakeTestConstants {
+    public final static int TEST_SCREENSHOTS_COUNT = 3;
+    public final static int TEST_VIDEOS_COUNT = 1;
+    public final static int TEST_LABELS_COUNT = 3;
+    public final static int TEST_ARGUMENTS_COUNT = 3;
+    public final static int TEST_STEPS_COUNT = 3;
+
+    public final static int STEP_SCREENSHOTS_COUNT = 3;
+    public final static int STEP_ARGUMENTS_COUNT = 3;
+
+    public final static String MANUAL_TEST_GUID = "some-manual-test-guid";
+}
