+++ testlio-mobile-java-testng-template/src/test/java/com/testlio/tests/login/TestlioLoginTest.java
@@ -1,4 +1,4 @@
-package com.testlio.tests;
+package com.testlio.tests.login;
 
 import com.testlio.lib.BaseTest;
 import com.testlio.pages.AlertPopup;
