+++ testlio-mobile-java-testng-template/src/test/java/com/testlio/tests/stress/ALotOfLabelsTest.java
