+++ testlio-mobile-java-testng-template/src/test/resources/suites/basic.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="Basic Suite">
+    <test name="When all tests in suite passed">
+        <classes>
+            <class name="com.testlio.tests.basic.AllPassedTest"/>
+        </classes>
+    </test>
+    <test name="When some tests in suite failed">
+        <classes>
+            <class name="com.testlio.tests.basic.SomeFailedTest"/>
+        </classes>
+    </test>
+    <test name="When some tests in suite errored">
+        <classes>
+            <class name="com.testlio.tests.basic.SomeErroredTest"/>
+        </classes>
+    </test>
+    <test name="When Testlio manual test ID defined">
+        <classes>
+            <class name="com.testlio.tests.basic.WithManualTestIdTest"/>
+        </classes>
+    </test>
+    <test name="When common problems arise">
+        <classes>
+            <class name="com.testlio.tests.basic.CommonProblemsTest"/>
+        </classes>
+    </test>
+</suite>
