--- testlio-web-java-testng-template/src/test/java/com/testlio/models/ResultType.java
@@ -1,14 +1,0 @@
-package com.testlio.models;
-
-import java.util.Random;
-
-public enum ResultType {
-     PASSED,
-     FAILED,
-     ERRORED;
-
-     public static ResultType getRandom() {
-          Random random = new Random();
-          return values()[random.nextInt(values().length)];
-     }
-}
